Remove multiplier 256 in some more places.

+++ src/server/target.h
@@ -28,8 +28,4 @@
 
 void Target_update(void);
 
-void Object_hits_target(
-    // move_state_t *ms,
-    object_t *obj,
-    target_t *targ,
-    long player_cost);
+void Object_hits_target(object_t *obj, target_t *targ, double player_cost);
